Move `RealTimeTripStateType` and `TransmodelRealTimeTripStateModel` to `timetable` package and improve test assertions
