+++ application/src/main/java/org/opentripplanner/apis/transmodel/model/timetable/TransmodelRealTimeTripStateModel.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.apis.transmodel.model.plan;
+package org.opentripplanner.apis.transmodel.model.timetable;
 
 /**
  * API model representing the real-time state of a service journey, used as the source object for
